storage: pull Replica batch routing into replica_send.go
Code movement only. Follows the pattern set in store_send.go.

Release note: None
